requirements.txt
django==5.0.7
djangorestframework==3.15.2
python-dotenv==1.0.1
django-model-utils==4.5.1
django-debug-toolbar==4.4.6
